Navigation, Prefabs and Simple AI
Took the navigation script from the tutorial I'm following and set up the nav mesh so the agent can move.

Added in Simple AI script that will likely be replaced later but for now functions to help test that the smart objects and navigation are working.
Changed 3D objects to prefabs so I could add an interaction point for the agent.