Added in Basic Needs
Continuing to follow the tutorial (video 2 now) Needs were added in.
Added in UI sliders above the character to visually show the needs.
Attached camera to character to make it easier to see needs sliders while testing.
Created a folder for TV scripts to declutter.
Created NotSoSimpleAI script and CommonAIBase script.
Moved code around in two previous scripts and SimpleAI script to make CommonAIBase the base script for the other two; adding the satisfying needs capability to both scripts.
At one point I thought the tutorial was wrong and managed to debug my way into fixing the issue I came across before realising I just missed the detail in the tutorial.
(Missed an "!" and the AI was only performing interactions it was not supposed to be able to perform)