Added Curves to Scores and Edited Traits
Added in a "HierarchyLevel" to the stats for the opportunity to give the stats different score curves based on the level they have, inspired by Moslow's Hierarchy of Needs.

Played around with traits and decreased agents pick size range from 5 to 2.
Traits are now "Well Rested", "Glutton", "Fun Chaser", and "Small Appetite".